deprecate backend part 7 of n
* Move around some code to kill the need for a backend method
* Type the backend arg to typing.Any since we just don't care now
